Removed NiceModal for new note
ref AP-1065

- New note dialog used the old design system's modal (NiceModal). This PR replaces it with Shade's Dialog component
